Restructured code: Shortened startup.py, moved view functions/login_manager to main.views, made mongodb singleton, used blueprint
