+++ templates/reference.html
@@ -61,9 +61,11 @@
 					<p>Convert XML files to real code</p>
 					<div class="row">
 						<div class="col-md-12 col-sm-12">
-	                        <form action="{{ url_for('logout') }}">				
+							{%endraw%}
+	                        <form action="{{ url_for('main_bp.logout') }}">				
 								<button class="btn btn-primary pull-right" id="log_out" ><span class="fa fa-user"></span>&nbsp Logout</button>
 							</form>
+							{%raw%}
 						</div>
 					</div>	
 				</div>
@@ -78,19 +80,19 @@
 					<!--Redirect to server status page here!-->
 					<li class="nav-item">
 						{%endraw%}
-						<form action="{{ url_for('serverstatus') }}" name="serverstatus" method="GET">
+						<form action="{{ url_for('main_bp.serverstatus') }}" name="serverstatus" method="GET">
 							<a class="nav-link" href="javascript: document.serverstatus.submit()">Server Status</a>
 						</form>
 					</li>
 					<!--Redirect to description page here!-->
 					<li class="nav-item">
-						<form action="{{ url_for('description') }}" name="description" method="GET">
-							<a class="nav-link" href="javascript: document.description.submit()">Server Status</a>
+						<form action="{{ url_for('main_bp.description') }}" name="description" method="GET">
+							<a class="nav-link" href="javascript: document.description.submit()">Description</a>
 						</form>
 					</li>
 					<!--Redirect to API reference page here!-->
 					<li class="nav-item">
-						<form action="{{ url_for('detail_instance') }}" name="detail_instance" method="GET">
+						<form action="{{ url_for('main_bp.detail_instance') }}" name="detail_instance" method="GET">
 						{%raw%}
 							<a class="nav-link" href="javascript: document.detail_instance.submit()">API Reference</a>
 						</form>
