+++ templates/result_page.html
@@ -75,7 +75,7 @@
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/semantic-ui/2.2.10/semantic.min.css">
     <script src="https://code.jquery.com/jquery-2.2.4.min.js"></script>
     <script type="text/javascript" src="{{ url_for('static', filename='js/actions.js') }}"></script>
-    <script type="text/javascript" src="{{ url_for('generated_code', path='default/' + db_name + '/Adapter.js') }}"></script>
+    <script type="text/javascript" src="{{ url_for('main_bp.generated_code', path='default/' + db_name + '/Adapter.js') }}"></script>
 
   </head>
   <body>
