+++ templates/xml_upload.html
@@ -54,7 +54,7 @@
 					<h4 align="center"><span class="fa fa-gear"></span>&nbsp Upload an XML that corressponds to your class diagram</h4>
 					<br>
 					<!--A form for user to submit for uploading the xml file-->
-					<form method="post" enctype="multipart/form-data" action="{{ url_for('result')}}">
+					<form method="post" enctype="multipart/form-data" action="{{ url_for('main_bp.result')}}">
 						<div class="container" style="padding:8px; border:1px solid black; border-radius:8px; width:50%">
 							<span class="fa fa-search"></span><input type="file" id="file" name="file" style="margin-left:5px">
 						</div>
